+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 1. Added composable to search weather
 2. Add user input to searh weather with latitude and longitude
 3. Add location entity for latitude and longitude
+4. Display weather data as grid
+    1. Add tailwind Css